% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mplus_engine.R
\name{mplus_engine}
\alias{mplus_engine}
\title{Mplus Engine for knitr}
\usage{
mplus_engine()
}
\value{
Used for side effects; registers knitr engine
}
\description{
Sets up a custom knitr engine for running Mplus code from Quarto/Rmarkdown documents.
}
\details{
When used in an mplus engine code block, this engine:
\enumerate{
\item Writes the code to formplus.inp
\item Executes Mplus on that file
\item Returns the output in the document
}
}
\examples{
\dontrun{
# Call mplus_engine() before knitting if you want to enable the engine.
}

}
\keyword{internal}
