% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_aliases.R
\name{pull_item_parameters}
\alias{pull_item_parameters}
\title{Pull Item Parameters from Mplus H5 Output (Preferred Name)}
\usage{
pull_item_parameters(h5, group1, group2, item, factorname)
}
\arguments{
\item{h5}{Character string: name of the h5 file}

\item{group1}{Character string: reference group label from Mplus output}

\item{group2}{Character string: focal group label from Mplus output}

\item{item}{Character string: item/variable name (continuous DV)}

\item{factorname}{Character string: name of the underlying latent variable}
}
\value{
A numeric vector c(a1, b1, a2, b2)
}
\description{
Preferred snake_case alias for pull_ItemParameters().
}
