% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_ItemParameters.R
\name{pull_ItemParameters}
\alias{pull_ItemParameters}
\title{Pull Item Parameters from Mplus H5 Output}
\usage{
pull_ItemParameters(h5, group1, group2, item, factorname)
}
\arguments{
\item{h5}{Character string: name of the h5 file}

\item{group1}{Character string: reference group label from Mplus output}

\item{group2}{Character string: focal group label from Mplus output}

\item{item}{Character string: item/variable name (continuous DV)}

\item{factorname}{Character string: name of the underlying latent variable}
}
\value{
A numeric vector c(a1, b1, a2, b2) where
a1 = intercept for group1, b1 = slope for group1,
a2 = intercept for group2, b2 = slope for group2
}
\description{
Extracts intercepts and slopes for an item from an H5RESULTS object
comparing two groups.
}
