% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_aliases.R
\name{pull_mu_sigma}
\alias{pull_mu_sigma}
\title{Pull Mean and Variance from Mplus H5 Output (Preferred Name)}
\usage{
pull_mu_sigma(h5, group2, factorname)
}
\arguments{
\item{h5}{Character string: name of the h5 file}

\item{group2}{Character string: focal group label from Mplus output}

\item{factorname}{Character string: name of the latent variable}
}
\value{
A numeric vector c(mu, sigma)
}
\description{
Preferred snake_case alias for pull_muSigma().
}
